Recombine pretty-print files to avoid cyclical deps
