--- src/tools/pretty/indent.ts
@@ -1,4 +1,0 @@
-/**
- * In its own file because otherwise there is a circular dependency issue.
- */
-export const INDENT = "   "
