--- src/tools/pretty/index.ts
@@ -1,5 +1,0 @@
-export * from "./indent"
-export * from "./print"
-export * from "./prompt"
-export * from "./spinners"
-export * from "./theme"
